[core] Split `plugins.py` into a `core/plugins` package
Break the monolithic `core/plugins.py` module into focused submodules
under `core/plugins`:

- `env.py`: ProcessEnvType, ProcessEnv, DirTreeProcessEnv,
  RezProcessEnv, and processEnvFactory (process-environment
  resolution for plugin execution)
- `base.py`: Plugin, NodePlugin, NodePluginStatus,
  validateNodeDesc, formatNodeDescriptionErrorMessage
- `manager.py`: NodePluginManager

Update all call sites to import from the new submodules.
