+++ meshroom/core/plugins/base.py
@@ -1,11 +1,9 @@
 from __future__ import annotations
 
-import glob
 import importlib
 import json
 import logging
 import os
-import re
 import sys
 
 from enum import Enum
@@ -15,8 +13,7 @@ from pathlib import Path
 from meshroom.common import BaseObject
 from meshroom.core import desc
 from meshroom.core.desc.attribute import ValueTypeErrors
-from meshroom import _MESHROOM_ROOT
-from meshroom.core.desc.node import _MESHROOM_COMPUTE_DEPS
+from meshroom.core.plugins.env import ProcessEnv
 
 
 def validateNodeDesc(nodeDesc: desc.BaseNode) -> list[tuple[str, ValueTypeErrors]]:
@@ -75,197 +72,6 @@ def formatNodeDescriptionErrorMessage(error: tuple[str, ValueTypeErrors]) -> str
     return f"Unknown error for parameter '{errMsg}'."
 
 
-class ProcessEnvType(Enum):
-    """ Supported process environments. """
-    DIRTREE = "dirtree",
-    REZ = "rez"
-
-
-class ProcessEnv(BaseObject):
-    """
-    Describes the environment required by a node's process.
-
-    Args:
-        folder: the source folder for the process.
-        configEnv: the dictionary containing the environment variables defined in a configuration file
-                   for the process to run.
-        pluginName: the name of the plugin object.
-        envType: (optional) the type of process environment.
-        uri: (optional) the Unique Resource Identifier to activate the environment.
-    """
-
-    def __init__(self, folder: str, configEnv: dict[str, str], pluginName: str,
-                 envType: ProcessEnvType = ProcessEnvType.DIRTREE, uri: str = ""):
-        super().__init__()
-        self._folder: str = folder
-        self._configEnv: dict[str: str] = configEnv
-        self.pluginName: str = pluginName
-        self._processEnvType: ProcessEnvType = envType
-        self.uri: str = uri
-        self._env: dict = None
-
-    def getEnvDict(self) -> dict:
-        """ Return the environment dictionary if it has been modified, None otherwise. """
-        return self._env
-
-    def getCommandPrefix(self) -> str:
-        """ Return the prefix to the command line that will be executed by the process. """
-        return ""
-
-    def getCommandSuffix(self) -> str:
-        """ Return the suffix to the command line that will be executed by the process. """
-        return ""
-
-
-class DirTreeProcessEnv(ProcessEnv):
-    """
-    """
-    def __init__(self, folder: str, configEnv: dict[str: str], pluginName: str):
-        super().__init__(folder, configEnv, pluginName, envType=ProcessEnvType.DIRTREE)
-
-        # If there is a virtual environment, it is expected to be named "venv".
-        # Beside the virtual environment, a standard "bin"/"lib"/"lib64" hierarchy at
-        # the top level of the plugin folder is expected.
-        venvFolder = Path(folder, "venv")
-
-        # Find all the libs that are not directly at the "lib*"-level
-        envLibPaths = glob.glob(f'{folder}/lib*/python[0-9].[0-9]*/site-packages',
-                                recursive=False)
-        venvLibPaths = glob.glob(f'{venvFolder}/lib*/python[0-9].[0-9]*/site-packages',
-                                 recursive=False)
-
-        self.binPaths: list = [str(Path(folder, "bin")), str(Path(venvFolder, "bin"))]
-        self.libPaths: list = [str(Path(folder, "lib")), str(Path(folder, "lib64")),
-                               str(Path(venvFolder, "lib")), str(Path(venvFolder, "lib64"))]
-        self.pythonPaths: list = [str(Path(folder)), str(Path(venvFolder))] + \
-                                 self.binPaths + envLibPaths + venvLibPaths
-
-        if sys.platform == "win32":
-            # For Windows platforms, try and include the content of the virtual env if it exists
-            # The virtual env is expected to be named "venv"
-            venvLibPath = Path(venvFolder, "Lib", "site-packages")
-            if venvLibPath.exists():
-                self.pythonPaths.append(venvLibPath.as_posix())
-        else:
-            # For Linux platforms, lib paths may need to be discovered recursively to be properly
-            # added to LD_LIBRARY_PATH
-            extraLibPaths = []
-            regex = re.compile(r"^lib(\d{2})?$")
-            for envPath in envLibPaths + venvLibPaths:
-                for path, directories, _ in os.walk(envPath):
-                    for directory in directories:
-                        if re.match(regex, directory):
-                            extraLibPaths.append(os.path.join(path, directory))
-            self.libPaths = self.libPaths + extraLibPaths
-
-        # Setup the environment dictionary
-        self._env = os.environ.copy()
-        self._env["PYTHONPATH"] = os.pathsep.join(
-            [f"{_MESHROOM_ROOT}"] + self.pythonPaths + [os.getenv('PYTHONPATH', '')])
-        self._env["LD_LIBRARY_PATH"] = f"{os.pathsep.join(self.libPaths)}{os.pathsep}{os.getenv('LD_LIBRARY_PATH', '')}"
-        self._env["PATH"] = f"{os.pathsep.join(self.binPaths)}{os.pathsep}{os.getenv('PATH', '')}"
-
-        for k, val in self._configEnv.items():
-            # Preserve user-defined environment variables:
-            # manually set environment variable values take precedence over config file defaults.
-            if k in self._env:
-                continue
-
-            self._env[k] = val
-
-
-class RezProcessEnv(ProcessEnv):
-    """
-    """
-
-    REZ_DELIMITER_PATTERN = re.compile(r"-|==|>=|>|<=|<")
-
-    def __init__(self, folder: str, configEnv: dict[str: str], pluginName: str, uri: str = ""):
-        if not uri:
-            raise RuntimeError("Missing name of the Rez environment needs to be provided.")
-        super().__init__(folder, configEnv, pluginName, envType=ProcessEnvType.REZ, uri=uri)
-
-    def resolveRezSubrequires(self) -> list[str]:
-        """
-        Return the list of packages defined for the node execution. These execution packages are
-        named subrequires.
-        Note: If a package does not have a version number, the version is aligned with the main
-        Meshroom environment (if this package is defined).
-        """
-        if os.getenv(f"{self.uri.upper()}_{self.pluginName.upper()}_SUBREQUIRES"):
-            subrequires = os.environ.get(f"{self.uri.upper()}_{self.pluginName.upper()}_SUBREQUIRES", "").split(os.pathsep)
-        else:
-            subrequires = os.environ.get(f"{self.uri.upper()}_SUBREQUIRES", "").split(os.pathsep)
-        if not subrequires:
-            return []
-
-        packages = []
-        # Packages that are resolved in the current environment
-        currentEnvPackages = []
-        resolvedVersions = {}
-        if "REZ_USED_RESOLVE" in os.environ:
-            resolvedPackages = os.getenv("REZ_USED_RESOLVE", "").split()
-            for package in resolvedPackages:
-                if package.startswith("~"):
-                    continue
-                currentEnvPackages.append(package)
-                name, version = self.REZ_DELIMITER_PATTERN.split(package, maxsplit=1)
-                resolvedVersions[name] = version
-        logging.debug("Packages in the current environment: " + ", ".join(currentEnvPackages))
-
-        # Take packages with the set versions for those which have one, and try to take packages
-        # in the current environment (if they are resolved in it)
-        for package in subrequires:
-            packageTuple = self.REZ_DELIMITER_PATTERN.split(package, maxsplit=1)
-            if len(packageTuple) == 1:
-                # Only the package name in the subrequires.
-                # Search for a corresponding version in the parent environment.
-                packageName = packageTuple[0]
-                parentResolvedVersion = resolvedVersions.get(packageName)
-                if parentResolvedVersion:
-                    packages.append(f"{packageName}=={parentResolvedVersion}")
-                else:
-                    packages.append(package)
-            elif len(packageTuple) == 2:
-                # The subrequires ask for a specific version
-                packages.append(package)
-
-        def extractPackageName(packageString: str) -> str:
-            return self.REZ_DELIMITER_PATTERN.split(packageString, maxsplit=1)[0]
-        packageNames = [extractPackageName(package) for package in packages]
-
-        for package in _MESHROOM_COMPUTE_DEPS:
-            # For packages that are required by meshroom_compute, do not specify any version
-            # or align it with Meshroom's: the version will be found during the resolution of
-            # the environment based on the other packages.
-            # If any of these packages is already part of the environment a plugin's dependency,
-            # do not add it
-            if package not in packageNames:
-                packages.append(package)
-
-        logging.debug("Packages for the execution environment: " + ", ".join(packages))
-        return packages
-
-    def getCommandPrefix(self):
-        # TODO: make Windows-compatible
-
-        # Use the PYTHONPATH from the subrequires' environment (which will only be resolved once
-        # inside the execution environment) and add MESHROOM_ROOT and the plugin's folder itself
-        # to it
-        pythonPaths = f"{os.pathsep.join(['$PYTHONPATH', f'{_MESHROOM_ROOT}', f'{self._folder}'])}"
-
-        return f"rez env {' '.join(self.resolveRezSubrequires())} -c 'PYTHONPATH={pythonPaths} "
-
-    def getCommandSuffix(self):
-        return "'"
-
-
-def processEnvFactory(folder: str, configEnv: dict[str: str], pluginName: str, envType: str = "dirtree", uri: str = "") -> ProcessEnv:
-    if envType == "dirtree":
-        return DirTreeProcessEnv(folder, configEnv, pluginName)
-    return RezProcessEnv(folder, configEnv, pluginName, uri=uri)
-
-
 class NodePluginStatus(Enum):
     """
     Loading status for NodePlugin objects.
@@ -294,7 +100,7 @@ class Plugin(BaseObject):
         configFullEnv: the static merge of os.environ and configEnv, with os.environ taking precedence
         processEnv: the environment required for the nodes' processes to be correctly executed
     """
-    
+
     _instancesCount = 0
 
     def __init__(self, name: str, path: str):
@@ -326,7 +132,7 @@ class Plugin(BaseObject):
     def name(self):
         """ Return the name of the plugin. """
         return self._name
-    
+
     @property
     def uname(self):
         """ Return the unique name of the plugin. """
@@ -619,181 +425,3 @@ class NodePlugin(BaseObject):
         if not self.plugin:
             return {}
         return self.plugin.configFullEnv
-
-class NodePluginManager(BaseObject):
-    """
-    Manager for all the loaded Plugin objects as well as the registered NodePlugin objects.
-
-    Members:
-        plugins: dictionary containing all the loaded Plugins, with their name as the key
-        nodePlugins: dictionary containing all the NodePlugins that have been registered
-                      (a NodePlugin may exist without having been registered) with their name as
-                      the key
-    """
-
-    def __init__(self):
-        super().__init__()
-
-        self._plugins: dict[str: Plugin] = {}  # loaded plugins
-        self._nodePlugins: dict[str: NodePlugin] = {}  # registered node plugins
-
-    def isRegistered(self, name: str) -> bool:
-        """
-        Return whether the node plugin has been registered already.
-
-        Args:
-            name: the name of the node plugin whose registration needs to be checked.
-        """
-        return name in self._nodePlugins
-
-    def belongsToPlugin(self, name: str) -> Plugin:
-        """
-        Check whether the node plugin belongs to a loaded plugin, independently from
-        whether it has been registered or not.
-
-        Args:
-            name: the name of the node plugin that needs to be searched for across plugins.
-
-        Returns:
-            Plugin | None: the Plugin the node belongs to if it exists, None otherwise.
-        """
-        for plugin in self._plugins.values():
-            if plugin.containsNodePlugin(name):
-                return plugin
-        return None
-
-    def getPlugins(self) -> dict[str: Plugin]:
-        """
-        Return a dictionary containing all the loaded Plugins, with {key, value} =
-        {name, Plugin}.
-        """
-        return self._plugins
-
-    def getPlugin(self, name: str, uname: bool = True) -> Plugin:
-        """
-        Return the loaded Plugin object with "name".
-
-        Args:
-            name: the unique name of the Plugin, used upon its loading.
-            uname: the name passed as argument is the unique name of the plugin.
-                   if set to False, we will search for any plugin with this name
-                   but this means there can be a collision. To avoid any confusion
-                   use this function with the unique name as much as possible.
-
-        Returns:
-            Plugin | None: the loaded Plugin object if it exists, None otherwise.
-        """
-        if uname:
-            # Find plugin with unique name
-            if name in self._plugins:
-                return self._plugins[name]
-        else:
-            for plugin in self._plugins.values():
-                if plugin.name == name:
-                    return plugin
-        return None
-
-    def addPlugin(self, plugin: Plugin, registerNodePlugins: bool = True):
-        """
-        Load a Plugin object.
-
-        Args:
-            plugin: the Plugin to load and add to the list of loaded plugins.
-            registerNodePlugins: True if all the NodePlugins from the plugin should be registered
-                                 at the same time the plugin is being loaded. Otherwise, the
-                                 NodePlugins will have to be registered at a later occasion.
-        """
-        pluginUName = plugin.uname
-        if self.getPlugin(pluginUName):
-            logging.warning(f"Plugin {pluginUName} is already registered.")
-            return
-        self._plugins[pluginUName] = plugin
-        if registerNodePlugins:
-            for node in plugin.nodes:
-                self.registerNode(plugin.nodes[node])
-
-    def removePlugin(self, plugin: Plugin, unregisterNodePlugins: bool = True):
-        """
-        Remove a loaded Plugin object.
-
-        Args:
-            plugin: the Plugin to remove from the list of loaded plugins.
-            unregisterNodePlugins: True if all the nodes from the plugin should be unregistered (if they
-                                   are registered) at the same time as the plugin is unloaded. Otherwise,
-                                   the registered NodePlugins will remain while the Plugin itself will
-                                   be unloaded.
-        """
-        if self.getPlugin(plugin.uname):
-            if unregisterNodePlugins:
-                for node in plugin.nodes.values():
-                    self.unregisterNode(node)
-            del self._plugins[plugin.uname]
-
-    def getRegisteredNodePlugins(self) -> dict[str: NodePlugin]:
-        """
-        Return a dictionary containing all the registered NodePlugins, with
-        {key, value} = {name, NodePlugin}.
-        """
-        return self._nodePlugins
-
-    def getRegisteredNodePlugin(self, name: str) -> NodePlugin:
-        """
-        Return the NodePlugin object that has been registered under the name "name" if it exists.
-
-        Args:
-            name: the name of the NodePlugin used for its registration.
-
-        Returns:
-            NodePlugin | None: the loaded NodePlugin object if it exists, None otherwise.
-        """
-        if self.isRegistered(name):
-            return self._nodePlugins[name]
-        return None
-
-    def registerNode(self, nodePlugin: NodePlugin):
-        """
-        Register a node plugin. A registered node plugin will become instantiable.
-        If it is already registered, or if there is an issue with the node description,
-        the node plugin will not be registered and its status will be updated.
-
-        Args:
-            nodePlugin: the node plugin to register.
-        """
-        name = nodePlugin.nodeDescriptor.__name__
-        if self.isRegistered(name):
-            existingPlugin: NodePlugin = self._nodePlugins[name]
-            logging.warning(
-                f"Could not register node {name} ({nodePlugin.path}) "
-                f"because another node is already registered with this name ({existingPlugin.path})"
-            )
-            return
-        if nodePlugin.status in (NodePluginStatus.DESC_ERROR,
-                                 NodePluginStatus.ERROR):
-            logging.warning(
-                f"Could not register node {name} ({nodePlugin.path}) "
-                f"because the node is in error ({nodePlugin.status})."
-            )
-            return
-
-        try:
-            self._nodePlugins[name] = nodePlugin
-            nodePlugin.status = NodePluginStatus.LOADED
-        except Exception as exc:
-            logging.error(f"NodePlugin {name} could not be loaded: {exc}")
-            nodePlugin.status = NodePluginStatus.LOADING_ERROR
-
-    def unregisterNode(self, nodePlugin: NodePlugin):
-        """
-        Unregister a node plugin. When unregistered, a node plugin cannot be instantiated anymore.
-        If it is not registered already, nothing happens.
-
-        Args:
-            nodePlugin: the node plugin to unregister.
-        """
-        name = nodePlugin.nodeDescriptor.__name__
-        if self.isRegistered(name):
-            if nodePlugin.status != NodePluginStatus.LOADED:
-                logging.warning(f"NodePlugin {name} is registered but is not correctly loaded.")
-            else:
-                nodePlugin.status = NodePluginStatus.NOT_LOADED
-            del self._nodePlugins[name]
