+++ tests/test_compatibility.py
@@ -8,7 +8,7 @@ from typing import Type
 import pytest
 
 from meshroom.core import desc, pluginManager
-from meshroom.core.plugins import NodePlugin
+from meshroom.core.plugins.base import NodePlugin
 from meshroom.core.exception import GraphCompatibilityError, NodeUpgradeError
 from meshroom.core.graph import Graph, loadGraph
 from meshroom.core.node import CompatibilityNode, CompatibilityIssue, Node
